Fixed tetrahedron in qlm & qlmACH, moved torque & force to multipoleLib and started torque tests
