+++ newt/qlmACH.py
@@ -316,13 +316,12 @@ def tetrahedron(L, rho, x, y, z):
     qlm[2, L+1] = -q00*3*z*(x-1j*y)/(4*np.sqrt(30))
     qlm[2, L+2] = q00*3*(x**2 - 1j*x*y - y**2)/(4*np.sqrt(30))
     qlm[3, L] = -q00*z*(x**2 + y**2 - 2*z**2)*np.sqrt(7)/40
-    #qlm[3, L+1] = q00*(3*x**3-1j*x**2*y+x*y**2-3j*y**3-4*(x-1j*y)*z**2)
-    qlm[3, L+1] = q00*(3*x**3-1j*x**2*y+x*y**2-3j*y**3-5*(x-1j*y)*z**2)
+    qlm[3, L+1] = q00*(3*x**3-1j*x**2*y+x*y**2-3j*y**3-4*(x-1j*y)*z**2)
     qlm[3, L+1] *= np.sqrt(7/3)/80
     qlm[3, L+2] = q00*z*(x**2-1j*x*y-y**2)*np.sqrt(7/30)/8
     qlm[3, L+3] = -q00*(x**2-y**2)*(x-1j*y)*np.sqrt(7/5)/16
-    qlm[4, L] = q00*(3*x**4+x**2*y**2+3*y**4-4*(x**2+y**2)*z**2+8*z**2)*3/280
-    q41 = z*(3*x**3-1j*x**2*y-2j*y**3+4j*y*z**2+x*(y**2-4*z**2))
+    qlm[4, L] = q00*(3*x**4+x**2*y**2+3*y**4-4*z**2*(x**2+y**2)+8*z**4)*3/280
+    q41 = z*(3*x**3-1j*x**2*y+x*y**2-3j*y**3-4*z**2*(x-1j*y))
     qlm[4, L+1] = q00*q41*3/(112*np.sqrt(5))
     q42 = 2*x**4-1j*x**3*y-1j*x*y**3-2*y**4-2*(x**2-1j*x*y-y**2)*z**2
     qlm[4, L+2] = -q00*q42*3/(56*np.sqrt(10))
@@ -330,7 +329,7 @@ def tetrahedron(L, rho, x, y, z):
     q44 = x**4-1j*x**3*y-x**2*y**2+1j*x*y**3+y**4
     qlm[4, L+4] = q00*q44*3/(8*np.sqrt(70))
     q50 = z*(3*x**4+x**2*y**2+3*y**4-4*(x**2+y**2)*z**2+8*z**4)
-    qlm[5, L] = q00*q50*np.sqrt(11)/4486
+    qlm[5, L] = q00*q50*np.sqrt(11)/448
     q51 = 5*x**5 - 1j*x**4*y + x**3*y**2 - 1j*x**2*y**3 + x*y**4 - 5j*y**5
     q51 += -2*(x+1j*y)*(3*x**2-4j*x*y-3*y**2)*z**2 + 8*(x-1j*y)*z**4
     qlm[5, L+1] = -q00*q51*np.sqrt(11/30)*3/448
@@ -339,7 +338,8 @@ def tetrahedron(L, rho, x, y, z):
     q53 = 5*x**5-3j*x**4*y-x**3*y**2-1j*x**2*y**3-3*x*y**4+5j*y**5
     q53 -= 4*(x**2-y**2)*(x-1j*y)*z**2
     qlm[5, L+3] = q00*q53*np.sqrt(11/35)/128
-    q54 = z**3*(x**4-1j*x**3*y-x**2*y**2+1j*x*y**3+y**4)
+    #q54 = z**3*(x**4-1j*x**3*y-x**2*y**2+1j*x*y**3+y**4)
+    q54 = z*(x**4-1j*x**3*y-x**2*y**2+1j*x*y**3+y**4)
     qlm[5, L+4] = q00*q54*np.sqrt(11/70)*3/64
     q55 = (x-1j*y)*((x**2-y**2)**2 + x**2*y**2)
     qlm[5, L+5] = -q00*q55*np.sqrt(11/7)*3/128
